Update
Validation, Notifications, Toasts, Fix

+++ www/app/Http/Controllers/LibraryController.php
@@ -2,43 +2,31 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StoreLibraryRequest;
+use App\Http\Requests\StoreRoleRequest;
 use App\Models\Audit_Log;
+use App\Models\Books;
 use App\Models\Library_Application;
 use App\Models\Role;
 use App\Models\User;
-use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Config;
-use Illuminate\Support\Facades\DB;
 use Illuminate\Http\Request;
 use App\Models\Library;
 use Illuminate\Support\Str;
-use JetBrains\PhpStorm\NoReturn;
-use Termwind\Components\Li;
 
 class LibraryController extends Controller
 {
 
-    public function storeLibrary(Request $request): \Illuminate\Routing\Redirector|\Illuminate\Http\RedirectResponse
+    public function storeLibrary(StoreLibraryRequest $request): \Illuminate\Http\JsonResponse
     {
-
         $userid = (Auth::id());
         $unique_key = Str::random(10);
-        $name = $request->get('name');
-
-        $image = $request->file('library_img');
-        if($image) {
-        $imgName = 'library_img_' . $name . '.png';
-            $image->move(public_path('/images/library'), $imgName);
-        }
-
 
         Library::create([
             'userid' => $userid,
-            'libraryName' => $name,
+            'libraryName' => $request->get('name'),
             'unique_key' => $unique_key,
             'organisation' => $request->get('organisation'),
-            'library_img' => $imgName ?? null,
             'description' => $request->get('description'),
         ]);
 
@@ -48,27 +36,7 @@ class LibraryController extends Controller
 
         Auth::login($user);
 
-        return redirect('/books');
-    }
-
-    public function globalSettings(Request $request): \Illuminate\Routing\Redirector|\Illuminate\Http\RedirectResponse
-    {
-        $name = $request->get('nameLibrary');
-        $image = $request->file('image');
-        $info = $request->get('info');
-
-        $imageName = 'image' . Library::count() + 1 . '.' . $image->getClientOriginalExtension();
-
-        $destinationPath = '/resources/images/library';
-        $image->move(public_path($destinationPath), $imageName);
-
-        Library::create([
-            'name' => $name,
-            'image' => $imageName,
-            'info' => $info,
-        ]);
-
-        return redirect('/books');
+        return response()->json(['redirect' => '/books']);
     }
 
     public function createRole(Request $request): \Illuminate\Contracts\View\View
@@ -78,12 +46,19 @@ class LibraryController extends Controller
         $roles = Role::where('unique_key', $unique_key)->get();
 
         return view('roles', compact('roles'));
-
     }
 
-    public function storeRole(Request $request): \Illuminate\Routing\Redirector|\Illuminate\Http\RedirectResponse
+    public function getRolesJson(Request $request): \Illuminate\Http\JsonResponse
+    {
+        $unique_key = Auth::user()->unique_key;
+        $roles = Role::where('unique_key', $unique_key)->get();
+        return response()->json(['data' => json_encode($roles)]);
+    }
+
+    public function storeRole(StoreRoleRequest $request): \Illuminate\Http\JsonResponse
     {
         $permissions = array_sum($request->input('permissions', []));
+
 
         Role::create([
             'name' => $request->input('name'),
@@ -91,13 +66,19 @@ class LibraryController extends Controller
             'unique_key' => Auth::user()->unique_key
         ]);
 
-        return redirect('/library/roles');
+        return response()->json(['message' => 'Роль успешно создана', 'status' => 'success']);
     }
 
-    public function deleteRole(Request $request, $id): \Illuminate\Routing\Redirector|\Illuminate\Http\RedirectResponse
+    public function deleteRole(Request $request, $id): \Illuminate\Http\JsonResponse
     {
+        $role = Role::find($id);
+
+        if (!$role) {
+            return response()->json(['status' => 'error', 'message' => 'Роль не найдена']);
+        }
+
         Role::destroy($id);
-        return redirect('/library/roles');
+        return response()->json(['status' => 'success', 'message' => 'Роль успешно удалена']);
     }
 
     public function assigningRole(Request $request): \Illuminate\Http\JsonResponse
@@ -105,13 +86,17 @@ class LibraryController extends Controller
         $role = Role::where('id', $request->get('idRole'))->first();
 
         $user = User::find($request->get('idUser'));
-        if ($role) {
-            $user->role = $role->name;
+        if($user) {
+            if ($role) {
+                $user->role = $role->name;
+            } else {
+                $user->role = null;
+            }
+            $user->save();
         } else {
-            $user->role = null;
+            return response()->json(['status' => 'error', 'message' => 'Пользователь не был найден']);
         }
-        $user->save();
-        return response()->json(true);
+        return response()->json(['status' => 'success', 'message' => 'Роль пользователя успешно изменена']);
     }
 
     public function allUsers(Request $request): \Illuminate\Contracts\View\View
@@ -136,9 +121,9 @@ class LibraryController extends Controller
 
         $library = Library::where('unique_key', $unique_key)->first();
         if ($library) {
-            return response()->json($library);
+            return response()->json(['message' => 'Библиотека успешно найдена', 'status' => 'success', 'library' => $library]);
         } else {
-            return response()->json(false);
+            return response()->json(['message' => 'Библиотека не найдена', 'status' => 'error']);
         }
     }
 
@@ -149,24 +134,37 @@ class LibraryController extends Controller
         $id = $request->input('id');
 
         $user = User::find($idUser);
+        if (!$user) {
+            return response()->json(['status' => 'error', 'message' => 'Данного пользователя не существует']);
+        }
+
+        $user = User::find($idUser);
         $user->unique_key = $unique_key;
         $user->save();
 
-        Library_Application::destroy($id);
+        $application = Library_Application::find($id);
+        if (!$application) {
+            return response()->json(['status' => 'error', 'message' => 'Заявка не найдена']);
+        }
 
-        return response()->json(true);
+        $application->delete();
+
+        return response()->json(['status' => 'success', 'message' => 'Заявка успешно принята']);
     }
 
     public function libraryDeleteApplication(Request $request): \Illuminate\Http\JsonResponse
     {
         $id = $request->input('id');
+        $application = Library_Application::find($id);
+        if (!$application) {
+            return response()->json(['status' => 'error', 'message' => 'Заявка не найдена']);
+        }
+        $application->delete();
 
-        Library_Application::destroy($id);
-
-        return response()->json(true);
+        return response()->json(['status' => 'success', 'message' => 'Заявка успешно отклонена']);
     }
 
-    public function libraryGetApplications(Request $request): \Illuminate\Http\JsonResponse
+    public function libraryGetApplications(Request $request): \Illuminate\Contracts\View\View
     {
         $unique_key = Auth::user()->unique_key;
 
@@ -176,7 +174,20 @@ class LibraryController extends Controller
         foreach ($applications as $application) {
             $users[] = User::find($application->idUser);
         }
+        return view('acceptApplications', compact('applications', 'users'));
 
+    }
+
+    public function libraryGetApplicationsJson(Request $request): \Illuminate\Http\JsonResponse
+    {
+        $unique_key = Auth::user()->unique_key;
+
+        $applications = Library_Application::where('unique_key', $unique_key)->get();
+
+        $users = [];
+        foreach ($applications as $application) {
+            $users[] = User::find($application->idUser);
+        }
         return response()->json(['applications' => $applications, 'users' => $users]);
     }
 
@@ -184,10 +195,14 @@ class LibraryController extends Controller
     {
         $unique_key = $request->input('code');
 
+        if(Library_Application::where('unique_key', $unique_key)->where('idUser', Auth::id())->exists()){
+            return response()->json(['message' => 'Вы уже отправляли заявку в данную библиотеку!', 'status' => 'error']);
+        }
+
         $library = Library::where('unique_key', $unique_key)->get();
 
         if (!$library) {
-            return false;
+            return response()->json(['message' => 'Ошибка при отправке заявки!', 'status' => 'error']);
         }
 
         Library_Application::create([
@@ -196,7 +211,9 @@ class LibraryController extends Controller
             'nameUser' => Auth::user()->name,
         ]);
 
-        return response()->json(true);
+
+
+        return response()->json(['message' => 'Заявка успешно отправлена', 'status' => 'success']);
     }
 
     public function deleteLibrary(Request $request): \Illuminate\Http\JsonResponse
@@ -212,12 +229,20 @@ class LibraryController extends Controller
         }
     }
 
-    public function allLogs(Request $request):\Illuminate\Contracts\View\View
+    public function allLogs(Request $request): \Illuminate\Contracts\View\View
     {
         $unique_key = Auth::user()->unique_key;
         $logs = Audit_Log::where('unique_key', $unique_key)->get();
 
         return view('libraryLogs', compact('logs'));
+    }
+
+    public function exitLibrary(Request $request): \Illuminate\Routing\Redirector|\Illuminate\Http\RedirectResponse
+    {
+        $user = User::find(Auth::id());
+        $user->unique_key = null;
+        $user->save();
+        return redirect('/');
     }
 
 }
